+++ src/main/java/ua/goit/dto/CustomersDTO.java
@@ -3,29 +3,29 @@ package ua.goit.dto;
 import java.util.List;
 
 public class CustomersDTO {
-    private long customerId;
+    private int customerId;
     private String customerName;
     private List<Long> projectIds;
 
     public CustomersDTO() {
     }
 
-    public CustomersDTO(long customerId, String customerName) {
+    public CustomersDTO(int customerId, String customerName) {
         this.customerId = customerId;
         this.customerName = customerName;
     }
 
-    public CustomersDTO(long customerId, String customerName, List<Long> projectIds) {
+    public CustomersDTO(int customerId, String customerName, List<Long> projectIds) {
         this.customerId = customerId;
         this.customerName = customerName;
         this.projectIds = projectIds;
     }
 
-    public long getCustomerId() {
+    public int getCustomerId() {
         return customerId;
     }
 
-    public void setCustomerId(long customerId) {
+    public void setCustomerId(int customerId) {
         this.customerId = customerId;
     }
